[feat] 제품 엔티티 생성
