--- src/main/java/com/example/musinsa_clone/domain/Delivery.java
@@ -1,30 +1,0 @@
-package com.example.musinsa_clone.domain;
-
-import com.example.musinsa_clone.domain.baseEntity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Delivery extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String province;
-
-    private String city;
-
-    private String town;
-
-    private String details;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-}
