+++ src/main/java/com/example/musinsa_clone/domain/enumClass/Gender.java
@@ -1,0 +1,5 @@
+package com.example.musinsa_clone.domain.enumClass;
+
+public enum Gender {
+    MALE, FEMALE
+}
