+++ src/main/java/com/example/musinsa_clone/domain/member/Member.java
@@ -1,5 +1,8 @@
-package com.example.musinsa_clone.domain;
+package com.example.musinsa_clone.domain.member;
 
+import com.example.musinsa_clone.domain.Coupon;
+import com.example.musinsa_clone.domain.enumClass.Gender;
+import com.example.musinsa_clone.domain.question.Question;
 import com.example.musinsa_clone.domain.baseEntity.BaseEntity;
 import com.example.musinsa_clone.domain.enumClass.Grade;
 import com.example.musinsa_clone.domain.enumClass.Role;
@@ -60,6 +63,10 @@ public class Member extends BaseEntity {
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private SocialType socialType;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Gender gender;
 
 
     // 매핑 테이블
